+++ src/main/java/com/example/ProyectoFinal/datos/PacienteDAO.java
@@ -1,6 +1,8 @@
 package com.example.ProyectoFinal.datos;
 
 import com.example.ProyectoFinal.datos.interfaces.IDao;
+import com.example.ProyectoFinal.exceptions.ElementAlreadyExistsException;
+import com.example.ProyectoFinal.exceptions.ResourceNotFoundException;
 import com.example.ProyectoFinal.model.Paciente;
 import com.example.ProyectoFinal.repository.IPacienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,24 +27,63 @@ public class PacienteDAO implements IDao<Paciente> {
     }
 
     @Override
-    public Paciente agregar(Paciente paciente) {
+    public Paciente agregar(Paciente paciente) throws ElementAlreadyExistsException {
+        if(pacienteRepository.findById(paciente.getId()).isEmpty())
+        {
+            throw new ElementAlreadyExistsException("El paciente a agregar ya existe en la base de datos.");
+        }
         paciente.setFechaAlta();
         return pacienteRepository.save(paciente);
     }
 
     @Override
-    public Paciente modificar(Paciente paciente) {
-        return pacienteRepository.save(paciente);
+    public Paciente modificar(Paciente paciente) throws ResourceNotFoundException {
+        if(pacienteRepository.findById(paciente.getId()).isEmpty())
+        {
+            throw new ResourceNotFoundException("El paciente a modificar no existe");
+        }
+        try
+        {
+            return pacienteRepository.save(paciente);
+        }
+        catch (Exception ex)
+        {
+            throw ex;
+        }
     }
 
     @Override
-    public Boolean eliminar(Long id) {
-        pacienteRepository.deleteById(id);
+    public Boolean eliminar(Long id) throws ResourceNotFoundException {
+
+        if(pacienteRepository.findById(id).isEmpty())
+        {
+            throw new ResourceNotFoundException("El paciente a eliminar no existe");
+        }
+        try
+        {
+            pacienteRepository.deleteById(id);
+        }
+        catch (Exception ex)
+        {
+            throw ex;
+        }
         return true;
     }
 
     @Override
-    public Paciente buscar(Long id) throws NoSuchElementException {
-        return pacienteRepository.findById(id).get();
+    public Paciente buscar(Long id) throws ResourceNotFoundException {
+
+        if(pacienteRepository.findById(id).isEmpty())
+        {
+            throw new ResourceNotFoundException("El paciente buscado no existe");
+        }
+        try
+        {
+            return pacienteRepository.findById(id).get();
+        }
+        catch (Exception ex)
+        {
+            throw ex;
+        }
     }
 }
